- Added navigation flow. - Fixed Activities null array issue. - Ui changes in LoginFragment.kt. - Improved Adapter code. - Changed model class variable type (Int -> Double).

+++ app/src/main/java/com/example/healthprotask/auth/model/UserActivitiesResponse.kt
@@ -15,7 +15,7 @@ data class UserActivitiesResponse(
         val activityParentName: String,
         val calories: Int,
         val description: String,
-        val distance: Int,
+        val distance: Double,
         val duration: Int,
         val hasActiveZoneMinutes: Boolean,
         val hasStartTime: Boolean,
@@ -50,7 +50,7 @@ data class UserActivitiesResponse(
     ) {
         data class Distance(
             val activity: String,
-            val distance: Int
+            val distance: Double
         )
     }
 }
